+++ minijinja/src/vm.rs
@@ -9,8 +9,8 @@ use crate::environment::Environment;
 use crate::error::{Error, ErrorKind};
 use crate::instructions::{Instruction, Instructions};
 use crate::key::Key;
-use crate::utils::{matches, RcType};
-use crate::value::{self, DynamicObject, Primitive, Value, ValueIterator};
+use crate::utils::matches;
+use crate::value::{self, Object, Primitive, RcType, Value, ValueIterator};
 use crate::AutoEscape;
 
 #[derive(Debug)]
@@ -19,8 +19,8 @@ pub struct LoopState {
     idx: AtomicUsize,
 }
 
-impl DynamicObject for LoopState {
-    fn fields(&self) -> &[&str] {
+impl Object for LoopState {
+    fn attributes(&self) -> &[&str] {
         &[
             "index0",
             "index",
@@ -150,7 +150,7 @@ impl<'env, 'context> Context<'env, 'context> {
                     ..
                 }) => {
                     if *with_loop_var && key == "loop" {
-                        return Some(Value::from_dynamic(controller.clone()));
+                        return Some(Value::from_rc_object(controller.clone()));
                     } else if let Some(value) = locals.get(key) {
                         return Some(value.clone());
                     }
@@ -165,8 +165,8 @@ impl<'env, 'context> Context<'env, 'context> {
                 }
             }
             if !cont {
-                if let Some(func) = env.get_function(key) {
-                    return Some(func.to_value());
+                if let Some(value) = env.get_global(key) {
+                    return Some(value);
                 }
                 break;
             }